+++ src/gpgpu-sim/l2cache.h
@@ -79,7 +79,7 @@ class memory_partition_unit {
 
   bool busy() const;
 
-  void cache_cycle(unsigned cycle);
+  void cache_cycle(unsigned long long cycle);
   void dram_cycle();
   void simple_dram_model_cycle();
 
@@ -169,7 +169,7 @@ class memory_sub_partition {
 
   bool busy() const;
 
-  void cache_cycle(unsigned cycle);
+  void cache_cycle(unsigned long long cycle);
 
   bool full() const;
   bool full(unsigned size) const;
